requirements.txt
pyobjc
emoji
webcolors